+++ app/src/main/kotlin/com/itrocket/union/selectCount/presentation/store/SelectCountResult.kt
@@ -2,6 +2,7 @@ package com.itrocket.union.selectCount.presentation.store
 
 import android.os.Parcelable
 import kotlinx.android.parcel.Parcelize
+import java.math.BigDecimal
 
 @Parcelize
-data class SelectCountResult(val id: String, val count: Long) : Parcelable
+data class SelectCountResult(val id: String, val count: BigDecimal) : Parcelable
