+++ union_sync_api/src/main/java/com/example/union_sync_api/entity/ReserveCountSyncEntity.kt
@@ -1,7 +1,9 @@
 package com.example.union_sync_api.entity
 
+import java.math.BigDecimal
+
 data class ReserveCountSyncEntity(
     val id: String,
-    val count: Long?,
+    val count: BigDecimal?,
     val userUpdated: String?
 )
