+++ union_sync_impl/src/main/java/com/example/union_sync_impl/entity/ActionRemainsRecordDb.kt
@@ -2,6 +2,7 @@ package com.example.union_sync_impl.entity
 
 import androidx.room.Entity
 import com.example.union_sync_impl.entity.core.SyncItemDb
+import java.math.BigDecimal
 
 @Entity(tableName = "action_remains_record")
 class ActionRemainsRecordDb(
@@ -9,7 +10,7 @@ class ActionRemainsRecordDb(
     cancel: Boolean? = false,
     val actionId: String,
     val remainId: String,
-    val count: Long?,
+    val count: BigDecimal?,
     updateDate: Long?,
     insertDate: Long?,
     userInserted: String?,
@@ -27,7 +28,7 @@ class ActionRemainsRecordDb(
         cancel: Boolean? = this.cancel,
         actionId: String = this.actionId,
         remainId: String = this.remainId,
-        count: Long? = this.count,
+        count: BigDecimal? = this.count,
         updateDate: Long? = this.updateDate,
         insertDate: Long? = this.insertDate,
         userInserted: String? = this.userInserted,
